requirements.txt
tensorflow>=2.12.0
pillow>=9.5.0
numpy>=1.24.0
flask>=2.3.0
python-dotenv>=1.0.0
flask-bootstrap>=3.3.7.1 